+++ lib/l10n/app_localizations_pt.dart
@@ -1600,4 +1600,8 @@ class AppLocalizationsPt extends AppLocalizations {
 
   @override
   String get download_retry => 'Tentar novamente';
+
+  @override
+  String get tv_cast_replay_hint =>
+      'Watched by casting from your phone. Send it again from your phone to play it here.';
 }